apply patch - refactor beanstandeteWahlbriefeGetter
